Added partial medal system Added ItemScripting (Item 243) Modified a few job advancement script, I'm going to redo everything. If anyone wants to help me get the npc text. Modified few MapleFamily stuff
More, but I'm too lazy to write them down.
Execute the new sql file btw



git-svn-id: https://svn6.assembla.com/svn/MoopleDEV@104 59296cf3-ab24-4890-a14e-c0dede3ece5d

+++ src/scripting/npc/NPCConversationManager.java
@@ -176,14 +176,14 @@ public class NPCConversationManager extends AbstractPlayerInteraction {
         return getPlayer().getJob().getId();
     }
 
-    public void startQuest(int id) {
+    public void startQuest(short id) {
         try {
             MapleQuest.getInstance(id).forceStart(getPlayer(), npc);
         } catch (NullPointerException ex) {
         }
     }
 
-    public void completeQuest(int id) {
+    public void completeQuest(short id) {
         try {
             MapleQuest.getInstance(id).forceComplete(getPlayer(), npc);
         } catch (NullPointerException ex) {
